GH-3698: Tidy up JSON-LD testing
