+++ jena-arq/src/test/java/org/apache/jena/riot/system/TS_RiotSystem.java
@@ -46,6 +46,7 @@ import org.junit.platform.suite.api.Suite;
 
     , TestRiotLib.class
     , TestAsyncParser.class
+    , TestJenaTitaniumConvert.class
 })
 
 /**
